[PATCH] remove size_t-based printk warnings
From: "Randy.Dunlap" <rddunlap@osdl.org>

This patch removes warnings on non-matching parameter types to printk
and incorrect function types (n_hdlc).

+++ fs/freevxfs/vxfs_inode.c
@@ -171,7 +171,7 @@ __vxfs_iget(ino_t ino, struct inode *ilistp)
 	return NULL;
 
 fail:
-	printk(KERN_WARNING "vxfs: unable to read inode %ld\n", ino);
+	printk(KERN_WARNING "vxfs: unable to read inode %ld\n", (unsigned long)ino);
 	vxfs_put_page(pp);
 	return NULL;
 }
